+++ client/src/hooks/README.tx
@@ -1,1 +1,1 @@
-Most likely temp, ill refactor for redux
+No longer temp, this will be the methd for authentication purposes, redux will be for different state.
